refactor: login 관련 type들 파일 분리
- types 파일 안에 있던 login 관련 type들을 별도의 파일에 넣음
