+++ src/api/types.ts
@@ -5,18 +5,6 @@ export interface IError {
   data: unknown | null;
 }
 
-export interface ILoginCode {
-  code: string;
-}
-
-export interface IKakaoAuthResponse {
-  kakaoId: number;
-  isMember: boolean;
-}
-
-export interface ILoginResponse {
-  accessToken: string;
-}
 
 export interface ICommentData {
   id: number;
