+++ src/api/types/login.ts
@@ -1,0 +1,12 @@
+export interface ILoginCode {
+  code: string;
+}
+
+export interface IKakaoAuthResponse {
+  kakaoId: number;
+  isMember: boolean;
+}
+
+export interface ILoginResponse {
+  accessToken: string;
+}
